refactor: move insert query building into query builder for testability.
